+++ src/modules/pubsub/pubsub-connection.ts
@@ -4,7 +4,8 @@ import {
   WebSocketConnection,
   TopicType,
   PubSubMessage,
-  NotifyEventType
+  NotifyEventType,
+  ParsedPubSubData
 } from './types.js'
 import * as WS from 'ws'
 
@@ -49,6 +50,14 @@ export class PubSubConnection implements WebSocketConnection {
     this.topics.push(...message.data.topics)
   }
 
+  listenToTopics(topics: string[]) {
+    const message = this.getListenMessageForTopic(topics)
+
+    this.sendMessage(message)
+
+    this.topics.push(...topics)
+  }
+
   sendMessage(message: PubSubMessage) {
     this.connection.send(JSON.stringify(message))
   }
@@ -58,12 +67,17 @@ export class PubSubConnection implements WebSocketConnection {
     channelId: number
   ): PubSubMessage => {
     const type = this.mapNotifyTypeToTopic(notifyType)
+
+    return this.getListenMessageForTopic([`${type}${channelId}`])
+  }
+
+  getListenMessageForTopic(topic: string[]): PubSubMessage {
     return {
       type: 'LISTEN',
       nonce: '',
       data: {
         auth_token: process.env.TWITCH_OAUTH,
-        topics: [`${type}${channelId}`]
+        topics: topic
       }
     }
   }
@@ -73,17 +87,26 @@ export class PubSubConnection implements WebSocketConnection {
     if (notifyType === NotifyEventType.STATUS) return TopicType.STATUS
   }
 
-  handleIncomingMessage({ data }: any) {
-    const parsedData = JSON.parse(data)
-    if (
-      parsedData.type !== 'RESPONSE' &&
-      parsedData.type !== 'MESSAGE' &&
-      parsedData.type !== 'PONG'
-    ) {
-      hb.log(LogType.PUBSUB, 'Other event ' + parsedData.type)
+  handleIncomingMessage({ data }: { data: string }) {
+    const parsedData: ParsedPubSubData = JSON.parse(data)
+
+    if (parsedData.type === 'RECONNECT') {
+      this.reconnect()
     }
 
     this.logError(parsedData.error)
+  }
+
+  async reconnect() {
+    this.connection.reconnect()
+    hb.log(
+      LogType.PUBSUB,
+      'A Pubsub connection has been closed and will restart'
+    )
+
+    this.listenToTopics(this.topics)
+
+    hb.log(LogType.PUBSUB, 'Connection successfully restartet')
   }
 
   logError(error: any) {