+++ CHANGELOG.md
@@ -16,6 +16,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Moved source files into src directory
 - updated database info text in readme 
+- join all pubsub topics at once
+
+### Fixed 
+
+- new connection after creating a new notification
 
 ## [1.1.0](https://github.com/helltf/helltfbot-v2/releases/tag/v1.1.0) - 2022-05-21